--- app/src/main/java/com/tuk/jetsetgo/presentation/addTravel/LoadingFragment.kt
@@ -1,15 +1,0 @@
-package com.tuk.jetsetgo.presentation.addTravel
-
-import com.tuk.jetsetgo.R
-import com.tuk.jetsetgo.databinding.FragmentAddTravelLoadingBinding
-import com.tuk.jetsetgo.presentation.base.BaseFragment
-
-class LoadingFragment: BaseFragment<FragmentAddTravelLoadingBinding>(R.layout.fragment_add_travel_loading) {
-    override fun initObserver() {
-
-    }
-
-    override fun initView() {
-
-    }
-}
