+++ app/src/main/java/com/tuk/jetsetgo/presentation/addTravel/TravelCountryFragment.kt
@@ -3,12 +3,12 @@ package com.tuk.jetsetgo.presentation.addTravel
 import android.view.View
 import androidx.navigation.fragment.findNavController
 import com.google.android.material.bottomnavigation.BottomNavigationView
-import com.tuk.jetsetgo.R
-import com.tuk.jetsetgo.databinding.FragmentAddTravel8Binding
 import com.tuk.jetsetgo.presentation.base.BaseFragment
 import com.tuk.jetsetgo.util.extension.setOnSingleClickListener
+import com.umc.jetsetgo.R
+import com.umc.jetsetgo.databinding.FragmentTravelCountryBinding
 
-class LocationFragment: BaseFragment<FragmentAddTravel8Binding>(R.layout.fragment_add_travel_8) {
+class TravelCountryFragment: BaseFragment<FragmentTravelCountryBinding>(R.layout.fragment_travel_country) {
     override fun initObserver() {
 
     }
@@ -42,8 +42,8 @@ class LocationFragment: BaseFragment<FragmentAddTravel8Binding>(R.layout.fragmen
     }
 
     private fun setupConfirmButton() {
-        binding.viewAddTravelSaveBtn.setOnSingleClickListener {
-            findNavController().navigate(R.id.goToLoading)
+        binding.viewConfirmBtn.setOnSingleClickListener {
+            findNavController().navigate(R.id.goToActivityLevel)
         }
     }
 }
