+++ app/src/main/java/com/tuk/jetsetgo/presentation/addTravel/TravelDateFragment.kt
@@ -3,12 +3,12 @@ package com.tuk.jetsetgo.presentation.addTravel
 import android.view.View
 import androidx.navigation.fragment.findNavController
 import com.google.android.material.bottomnavigation.BottomNavigationView
-import com.tuk.jetsetgo.R
-import com.tuk.jetsetgo.databinding.FragmentAddTravel6Binding
 import com.tuk.jetsetgo.presentation.base.BaseFragment
 import com.tuk.jetsetgo.util.extension.setOnSingleClickListener
+import com.umc.jetsetgo.R
+import com.umc.jetsetgo.databinding.FragmentTravelDateBinding
 
-class BudgetFragment: BaseFragment<FragmentAddTravel6Binding>(R.layout.fragment_add_travel_6) {
+class TravelDateFragment: BaseFragment<FragmentTravelDateBinding>(R.layout.fragment_travel_date) {
     override fun initObserver() {
 
     }
@@ -43,7 +43,7 @@ class BudgetFragment: BaseFragment<FragmentAddTravel6Binding>(R.layout.fragment_
 
     private fun setupConfirmButton() {
         binding.viewConfirmBtn.setOnSingleClickListener {
-            findNavController().navigate(R.id.goToDate)
+            findNavController().navigate(R.id.goToLocation)
         }
     }
 }
