+++ app/src/main/java/com/tuk/jetsetgo/presentation/addTravel/TravelLoadingFragment.kt
@@ -1,0 +1,15 @@
+package com.tuk.jetsetgo.presentation.addTravel
+
+import com.tuk.jetsetgo.presentation.base.BaseFragment
+import com.umc.jetsetgo.R
+import com.umc.jetsetgo.databinding.FragmentTravelLoadingBinding
+
+class TravelLoadingFragment: BaseFragment<FragmentTravelLoadingBinding>(R.layout.fragment_travel_loading) {
+    override fun initObserver() {
+
+    }
+
+    override fun initView() {
+
+    }
+}
