+++ app/src/main/java/com/tuk/jetsetgo/presentation/addTravel/TravelLocationFragment.kt
@@ -3,12 +3,12 @@ package com.tuk.jetsetgo.presentation.addTravel
 import android.view.View
 import androidx.navigation.fragment.findNavController
 import com.google.android.material.bottomnavigation.BottomNavigationView
-import com.tuk.jetsetgo.R
-import com.tuk.jetsetgo.databinding.FragmentAddTravel4Binding
 import com.tuk.jetsetgo.presentation.base.BaseFragment
 import com.tuk.jetsetgo.util.extension.setOnSingleClickListener
+import com.umc.jetsetgo.R
+import com.umc.jetsetgo.databinding.FragmentTravelLocationBinding
 
-class PurposeFragment: BaseFragment<FragmentAddTravel4Binding>(R.layout.fragment_add_travel_4) {
+class TravelLocationFragment: BaseFragment<FragmentTravelLocationBinding>(R.layout.fragment_travel_location) {
     override fun initObserver() {
 
     }
@@ -42,8 +42,8 @@ class PurposeFragment: BaseFragment<FragmentAddTravel4Binding>(R.layout.fragment
     }
 
     private fun setupConfirmButton() {
-        binding.viewConfirmBtn.setOnSingleClickListener {
-            findNavController().navigate(R.id.goToTheme)
+        binding.viewAddTravelSaveBtn.setOnSingleClickListener {
+            findNavController().navigate(R.id.goToLoading)
         }
     }
 }
