+++ app/src/main/java/com/tuk/jetsetgo/presentation/addTravel/TravelThemeFragment.kt
@@ -3,12 +3,12 @@ package com.tuk.jetsetgo.presentation.addTravel
 import android.view.View
 import androidx.navigation.fragment.findNavController
 import com.google.android.material.bottomnavigation.BottomNavigationView
-import com.tuk.jetsetgo.R
-import com.tuk.jetsetgo.databinding.FragmentAddTravel3Binding
 import com.tuk.jetsetgo.presentation.base.BaseFragment
 import com.tuk.jetsetgo.util.extension.setOnSingleClickListener
+import com.umc.jetsetgo.R
+import com.umc.jetsetgo.databinding.FragmentTravelThemeBinding
 
-class ActivityLevelFragment: BaseFragment<FragmentAddTravel3Binding>(R.layout.fragment_add_travel_3) {
+class TravelThemeFragment:BaseFragment<FragmentTravelThemeBinding>(R.layout.fragment_travel_theme) {
     override fun initObserver() {
 
     }
@@ -43,7 +43,7 @@ class ActivityLevelFragment: BaseFragment<FragmentAddTravel3Binding>(R.layout.fr
 
     private fun setupConfirmButton() {
         binding.viewConfirmBtn.setOnSingleClickListener {
-            findNavController().navigate(R.id.goToPurpose)
+            findNavController().navigate(R.id.goToBudget)
         }
     }
 }
