--- app/src/main/res/layout/fragment_add_travel_6.xml
@@ -1,77 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <data>
-
-    </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="여행 예산을\n입력해주세요"
-            style="@style/TextAppearance.JSG.Head1"
-            android:lineSpacingExtra="5dp"
-            android:layout_marginStart="30dp"
-            android:layout_marginTop="150dp"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"/>
-
-        <TextView
-            android:id="@+id/tv_addTravel_budget"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="여행 예산"
-            style="@style/TextAppearance.JSG.Body1"
-            android:layout_marginBottom="10dp"
-            android:layout_marginStart="30dp"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintBottom_toTopOf="@id/tv_addTravel_dropDown_budget"
-            />
-
-        <EditText
-            android:id="@+id/tv_addTravel_dropDown_budget"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:hint="예산을 입력해주세요"
-            style="@style/TextAppearance.JSG.Body1"
-            android:inputType="number"
-            android:dropDownHeight="wrap_content"
-            android:dropDownWidth="wrap_content"
-            android:layout_marginStart="30dp"
-            android:layout_marginBottom="50dp"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toStartOf="@id/tv_addTravel_won"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintBottom_toBottomOf="parent"/>
-
-        <TextView
-            android:id="@+id/tv_addTravel_won"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="₩"
-            style="@style/TextAppearance.JSG.Body1"
-            android:layout_marginEnd="30dp"
-            app:layout_constraintStart_toEndOf="@id/tv_addTravel_dropDown_budget"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toTopOf="@id/tv_addTravel_dropDown_budget"
-            app:layout_constraintBottom_toBottomOf="@id/tv_addTravel_dropDown_budget"
-            />
-
-        <View
-            android:id="@+id/view_confirm_btn"
-            android:layout_width="0dp"
-            android:layout_height="0dp"
-            android:background="@drawable/btn_next"
-            app:layout_constraintDimensionRatio="1:1"
-            app:layout_constraintWidth_percent="0.15"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintBottom_toBottomOf="parent"
-            android:layout_margin="30dp"/>
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
