+++ app/src/main/res/layout/fragment_travel_date.xml
@@ -13,7 +13,7 @@
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="활동량을\n선택해주세요"
+            android:text="여행 날짜를\n선택해주세요"
             style="@style/TextAppearance.JSG.Head1"
             android:lineSpacingExtra="5dp"
             android:layout_marginStart="30dp"
@@ -22,22 +22,22 @@
             app:layout_constraintTop_toTopOf="parent"/>
 
         <TextView
-            android:id="@+id/tv_addTravel_visitCount"
+            android:id="@+id/tv_travelDate_startDate"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="하루에 방문할 관광지 개수"
+            android:text="시작 날짜"
             style="@style/TextAppearance.JSG.Body1"
             android:layout_marginBottom="10dp"
             android:layout_marginStart="30dp"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintBottom_toTopOf="@id/tv_addTravel_dropDown_visitCount"
+            app:layout_constraintBottom_toTopOf="@id/tv_travelDate_dropDown_startDate"
             />
 
         <AutoCompleteTextView
-            android:id="@+id/tv_addTravel_dropDown_visitCount"
+            android:id="@+id/tv_travelDate_dropDown_startDate"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
-            android:hint="개수를 선택해주세요"
+            android:hint="시작 날짜를 선택해주세요"
             style="@style/TextAppearance.JSG.Body1"
             android:inputType="none"
             android:dropDownHeight="wrap_content"
@@ -50,59 +50,31 @@
             app:layout_constraintBottom_toBottomOf="parent"/>
 
         <TextView
-            android:id="@+id/tv_addTravel_time"
+            android:id="@+id/tv_travelDate_endDate"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="활동 시작 시간 / 종료 시간"
+            android:text="종료 날짜"
             style="@style/TextAppearance.JSG.Body1"
             android:layout_marginTop="30dp"
             android:layout_marginBottom="10dp"
             android:layout_marginStart="30dp"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/tv_addTravel_dropDown_visitCount"
+            app:layout_constraintTop_toBottomOf="@id/tv_travelDate_dropDown_startDate"
             />
 
         <AutoCompleteTextView
-            android:id="@+id/tv_addTravel_dropDown_start"
+            android:id="@+id/tv_travelDate_dropDown_endDate"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
-            android:hint="시작 시간"
+            android:hint="종료 날짜를 선택해주세요"
             style="@style/TextAppearance.JSG.Body1"
             android:inputType="none"
             android:dropDownHeight="wrap_content"
             android:dropDownWidth="wrap_content"
             android:layout_marginHorizontal="30dp"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toStartOf="@id/tv_addTravel_text1"
-            app:layout_constraintTop_toBottomOf="@id/tv_addTravel_time"
-            android:gravity="center"/>
-
-        <TextView
-            android:id="@+id/tv_addTravel_text1"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="~"
-            style="@style/TextAppearance.JSG.Body1"
-            app:layout_constraintStart_toEndOf="@id/tv_addTravel_dropDown_start"
-            app:layout_constraintEnd_toEndOf="@id/tv_addTravel_dropDown_end"
-            app:layout_constraintTop_toTopOf="@id/tv_addTravel_dropDown_start"
-            app:layout_constraintBottom_toBottomOf="@id/tv_addTravel_dropDown_end"
-            />
-
-        <AutoCompleteTextView
-            android:id="@+id/tv_addTravel_dropDown_end"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:hint="종료 시간"
-            style="@style/TextAppearance.JSG.Body1"
-            android:inputType="none"
-            android:dropDownHeight="wrap_content"
-            android:dropDownWidth="wrap_content"
-            android:layout_marginHorizontal="30dp"
-            app:layout_constraintStart_toEndOf="@id/tv_addTravel_text1"
             app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/tv_addTravel_time"
-            android:gravity="center"/>
+            app:layout_constraintTop_toBottomOf="@id/tv_travelDate_endDate"/>
 
         <View
             android:id="@+id/view_confirm_btn"
