+++ app/src/main/res/layout/fragment_travel_loading.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <data>
 
@@ -13,7 +12,7 @@
         android:background="@color/blue_main">
 
         <TextView
-            android:id="@+id/tv_addTravel_head1"
+            android:id="@+id/tv_travelLoading_head1"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="계획을 생성 중입니다..."
@@ -23,5 +22,15 @@
             app:layout_constraintTop_toTopOf="parent"
             app:layout_constraintBottom_toBottomOf="parent"/>
 
+        <View
+            android:id="@+id/view_travelLoading_home_btn"
+            android:layout_width="0dp"
+            android:layout_height="0dp"
+            android:background="@drawable/ic_unselected_home"
+            app:layout_constraintDimensionRatio="1:1"
+            app:layout_constraintWidth_percent="0.15"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintBottom_toBottomOf="parent"
+            android:layout_margin="30dp"/>
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
