+++ app/src/main/res/layout/fragment_travel_location.xml
@@ -1,0 +1,62 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools">
+
+    <data>
+
+    </data>
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent">
+
+        <TextView
+            android:id="@+id/tv_travelLocation_head1"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="원하는 방문지를\n골라볼까요?"
+            style="@style/TextAppearance.JSG.Head1"
+            android:lineSpacingExtra="5dp"
+            android:layout_marginStart="30dp"
+            android:layout_marginTop="150dp"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent"/>
+
+        <androidx.recyclerview.widget.RecyclerView
+            android:layout_width="match_parent"
+            android:layout_height="0dp"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/tv_travelLocation_head1"
+            app:layout_constraintBottom_toTopOf="@id/view_travelLocation_add_btn"
+            android:layout_marginTop="50dp"
+            android:layout_marginBottom="30dp"
+            android:layout_marginHorizontal="30dp"
+            tools:listitem="@layout/item_location"
+            android:orientation="vertical"/>
+
+        <View
+            android:id="@+id/view_travelLocation_add_btn"
+            android:layout_width="0dp"
+            android:layout_height="0dp"
+            android:background="@drawable/btn_next"
+            app:layout_constraintDimensionRatio="1:1"
+            app:layout_constraintWidth_percent="0.15"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintBottom_toTopOf="@id/view_travelLocation_save_btn"
+            android:layout_marginHorizontal="30dp"
+            android:layout_marginVertical="20dp"/>
+
+        <View
+            android:id="@+id/view_travelLocation_save_btn"
+            android:layout_width="0dp"
+            android:layout_height="0dp"
+            android:background="@drawable/btn_next"
+            app:layout_constraintDimensionRatio="1:1"
+            app:layout_constraintWidth_percent="0.15"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintBottom_toBottomOf="parent"
+            android:layout_margin="30dp"/>
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
+</layout>
