+++ app/src/main/res/navigation/main_graph.xml
@@ -5,104 +5,110 @@
     android:id="@+id/main_graph"
     app:startDestination="@id/homeFragment">
 
+    <!-- home -->
     <fragment
         android:id="@+id/homeFragment"
         android:name="com.tuk.jetsetgo.presentation.home.HomeFragment"
         android:label="HomeFragment"
         tools:layout="@layout/fragment_home"/>
+
+    <!-- review -->
     <fragment
         android:id="@+id/reviewFragment"
         android:name="com.tuk.jetsetgo.presentation.review.ReviewFragment"
         android:label="ReviewFragment"
         tools:layout="@layout/fragment_review"/>
 
+    <!-- add travel -->
     <navigation android:id="@+id/addTravelTab"
         app:startDestination="@+id/addTravelFragment">
         <fragment
-            android:id="@+id/addTravelFragment"
-            android:name="com.tuk.jetsetgo.presentation.addTravel.AddTravelFragment"
+            android:id="@+id/personnelFragment"
+            android:name="com.tuk.jetsetgo.presentation.addTravel.TravelPersonnelFragment"
             android:label="AddTravelFragment"
-            tools:layout="@layout/fragment_add_travel">
+            tools:layout="@layout/fragment_travel_personnel">
             <action
                 android:id="@+id/goToCountry"
-                app:destination="@id/selectCountryFragment" />
+                app:destination="@id/personnelFragment" />
         </fragment>
         <fragment
-            android:id="@+id/selectCountryFragment"
-            android:name="com.tuk.jetsetgo.presentation.addTravel.SelectCountryFragment"
+            android:id="@+id/countryFragment"
+            android:name="com.tuk.jetsetgo.presentation.addTravel.TravelCountryFragment"
             android:label="SelectCountryFragment"
-            tools:layout="@layout/fragment_add_travel_2">
+            tools:layout="@layout/fragment_travel_country">
             <action
                 android:id="@+id/goToActivityLevel"
                 app:destination="@id/activityLevelFragment" />
         </fragment>
         <fragment
             android:id="@+id/activityLevelFragment"
-            android:name="com.tuk.jetsetgo.presentation.addTravel.ActivityLevelFragment"
+            android:name="com.tuk.jetsetgo.presentation.addTravel.TravelActivityLevelFragment"
             android:label="ActivityLevelFragment"
-            tools:layout="@layout/fragment_add_travel_3">
+            tools:layout="@layout/fragment_travel_activity_level">
             <action
                 android:id="@+id/goToPurpose"
                 app:destination="@id/purposeFragment" />
         </fragment>
         <fragment
             android:id="@+id/purposeFragment"
-            android:name="com.tuk.jetsetgo.presentation.addTravel.PurposeFragment"
+            android:name="com.tuk.jetsetgo.presentation.addTravel.TravelPurposeFragment"
             android:label="PurposeFragment"
-            tools:layout="@layout/fragment_add_travel_4">
+            tools:layout="@layout/fragment_travel_purpose">
             <action
                 android:id="@+id/goToTheme"
                 app:destination="@id/themeFragment" />
         </fragment>
         <fragment
             android:id="@+id/themeFragment"
-            android:name="com.tuk.jetsetgo.presentation.addTravel.ThemeFragment"
+            android:name="com.tuk.jetsetgo.presentation.addTravel.TravelThemeFragment"
             android:label="ThemeFragment"
-            tools:layout="@layout/fragment_add_travel_5">
+            tools:layout="@layout/fragment_travel_theme">
             <action
                 android:id="@+id/goToBudget"
                 app:destination="@id/budgetFragment" />
         </fragment>
         <fragment
             android:id="@+id/budgetFragment"
-            android:name="com.tuk.jetsetgo.presentation.addTravel.BudgetFragment"
+            android:name="com.tuk.jetsetgo.presentation.addTravel.TravelBudgetFragment"
             android:label="BudgetFragment"
-            tools:layout="@layout/fragment_add_travel_6">
+            tools:layout="@layout/fragment_travel_budget">
             <action
                 android:id="@+id/goToDate"
                 app:destination="@id/dateFragment" />
         </fragment>
         <fragment
             android:id="@+id/dateFragment"
-            android:name="com.tuk.jetsetgo.presentation.addTravel.DateFragment"
+            android:name="com.tuk.jetsetgo.presentation.addTravel.TravelDateFragment"
             android:label="DateFragment"
-            tools:layout="@layout/fragment_add_travel_7">
+            tools:layout="@layout/fragment_travel_date">
             <action
                 android:id="@+id/goToLocation"
                 app:destination="@id/locationFragment" />
         </fragment>
         <fragment
             android:id="@+id/locationFragment"
-            android:name="com.tuk.jetsetgo.presentation.addTravel.LocationFragment"
+            android:name="com.tuk.jetsetgo.presentation.addTravel.TravelLocationFragment"
             android:label="LocationFragment"
-            tools:layout="@layout/fragment_add_travel_8">
+            tools:layout="@layout/fragment_travel_location">
             <action
                 android:id="@+id/goToLoading"
                 app:destination="@id/loadingFragment" />
         </fragment>
         <fragment
             android:id="@+id/loadingFragment"
-            android:name="com.tuk.jetsetgo.presentation.addTravel.LoadingFragment"
+            android:name="com.tuk.jetsetgo.presentation.addTravel.TravelLoadingFragment"
             android:label="LoadingFragment"
-            tools:layout="@layout/fragment_add_travel_loading"/>
-
+            tools:layout="@layout/fragment_travel_loading"/>
     </navigation>
 
+    <!-- my Travel -->
     <fragment
         android:id="@+id/myTravelFragment"
         android:name="com.tuk.jetsetgo.presentation.myTravel.MyTravelFragment"
         android:label="MyTravelFragment"
         tools:layout="@layout/fragment_my_travel"/>
+
+    <!-- my page -->
     <fragment
         android:id="@+id/myPageFragment"
         android:name="com.tuk.jetsetgo.presentation.mypage.MypageFragment"
